+++ packages/whale-api-client/src/index.ts
@@ -1,11 +1,11 @@
 export * from './errors'
 
-export * as address from './api/address'
-export * as poolpair from './api/poolpair'
 export * as rpc from './api/rpc'
+export * as address from './api/address'
+export * as poolpairs from './api/poolpairs'
 export * as transactions from './api/transactions'
 export * as tokens from './api/tokens'
-export * as masternodes from './api/masternode'
+export * as masternodes from './api/masternodes'
 export * as blocks from './api/blocks'
 
 export * from './whale.api.client'
